feature(feedback): separate middlewares from other functions
Starts #149862942
