+++ server/routes/index.js
@@ -2,53 +2,55 @@ import rolesController from '../controllers/roles';
 import usersController from '../controllers/users';
 import documentsController from '../controllers/documents';
 import searchController from '../controllers/search';
-import Utils from '../helper/utils';
-import utils from '../helper/documentsLogic';
+import Middleware from '../middleware/middleware';
 
 const Route = (app) => {
   // Endpoints for roles
   app.post('/api/v1/roles',
-    Utils.isLoggedIn, Utils.isSuper, rolesController.create);
+    Middleware.authenticate, Middleware.isSuper, rolesController.create);
   app.get('/api/v1/roles',
-    Utils.isLoggedIn, Utils.isAdmin, rolesController.list);
+    Middleware.authenticate, Middleware.isAdmin, rolesController.list);
 
   // Endpoints for users
-  app.post('/api/v1/users/login', Utils.isValid, usersController.login);
-  app.post('/api/v1/users', Utils.isValid, usersController.create);
+  app.post('/api/v1/users/login',
+    Middleware.isUserValid, usersController.login);
+  app.post('/api/v1/users', Middleware.isUserValid, usersController.create);
 
-  app.get('/api/v1/users/logout', Utils.isLoggedIn, usersController.logout);
+  app.get('/api/v1/users/logout',
+    Middleware.authenticate, usersController.logout);
   app.get('/api/v1/users',
-    Utils.isLoggedIn, Utils.isAdmin, usersController.list);
+    Middleware.authenticate, Middleware.isAdmin, usersController.list);
   app.get('/api/v1/users/:userId',
-    Utils.isLoggedIn, Utils.isAdmin, usersController.retrieve);
+    Middleware.authenticate, Middleware.isAdmin, usersController.retrieve);
   app.get('/api/v1/users/:userId/documents',
-    Utils.isLoggedIn, Utils.isAdmin, usersController.allUsersDoc);
+    Middleware.authenticate, Middleware.isAdmin, usersController.allUsersDoc);
 
-  app.put('/api/v1/users/:userId', Utils.isLoggedIn, usersController.update);
+  app.put('/api/v1/users/:userId',
+    Middleware.authenticate, usersController.update);
 
   app.delete('/api/v1/users/:userId',
-    Utils.isLoggedIn, Utils.isSuper, usersController.destroy);
+    Middleware.authenticate, Middleware.isSuper, usersController.destroy);
 
   // Endpoints for documents
-  app.post('/api/v1/documents',
-    Utils.isLoggedIn, utils.isValid, documentsController.create);
+  app.post('/api/v1/documents', Middleware.authenticate,
+    Middleware.isDocumentValid, documentsController.create);
 
   app.get('/api/v1/documents',
-    Utils.isLoggedIn, documentsController.list);
+    Middleware.authenticate, documentsController.list);
   app.get('/api/v1/documents/:docId',
-    Utils.isLoggedIn, documentsController.retrieve);
+    Middleware.authenticate, documentsController.retrieve);
 
   app.put('/api/v1/documents/:docId',
-    Utils.isLoggedIn, documentsController.update);
+    Middleware.authenticate, documentsController.update);
 
   app.delete('/api/v1/documents/:docId',
-    Utils.isLoggedIn, documentsController.destroy);
+    Middleware.authenticate, documentsController.destroy);
 
   // Endpoints for Search
   app.get('/api/v1/search/users',
-    Utils.isLoggedIn, Utils.isAdmin, searchController.userSearch);
+    Middleware.authenticate, Middleware.isAdmin, searchController.userSearch);
 
   app.get('/api/v1/search/documents',
-    Utils.isLoggedIn, searchController.docSearch);
+    Middleware.authenticate, searchController.docSearch);
 };
 export default Route;
